% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGM.r
\name{make_MGM_list}
\alias{make_MGM_list}
\title{Make MGM lists from input data}
\usage{
make_MGM_list(X, Y, group)
}
\arguments{
\item{X}{data frame or matrix of continuous variables (row: observation, column: variable)}

\item{Y}{data frame or matrix of discrete variables (row: observation, column: variable)}

\item{group}{group variable vector, with the sample names}
}
\description{
Make MGM lists from input data
}
